pyproject.toml
[tool.poetry]
name = "router-lab"
version = "0.1.0"
description = "Make your own routing network!"
authors = ["wldhg <wldhg@aru.qa>"]
readme = "README.md"
packages = [{ include = "router_lab" }]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.8.0"
numpy = "^1.24.3"
loguru = "^0.7.0"
typer = "^0.9.0"
rich = "^13.3.5"
aiohttp = "^3.8.4"
aiohttp-cors = "^0.7.0"
typeguard = "^4.0.0"

[tool.poetry.group.dev.dependencies]
bump2version = "^1.0.1"
black = "^23.3.0"
pyright = "^1.1.309"

[tool.poetry.scripts]
rlab = "router_lab.__main__:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.pyright]
include = ["router_lab"]
pythonVersion = "3.10"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
reportPropertyTypeMismatch = "error"
reportFunctionMemberAccess = "error"
reportDuplicateImport = "error"
reportPrivateUsage = "error"
reportTypeCommentUsage = "error"
reportConstantRedefinition = "error"
reportMissingParameterType = "warning"
reportImplicitStringConcatenation = "warning"
reportUnnecessaryIsInstance = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryContains = "error"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnusedClass = "warning"
reportUnusedFunction = "warning"
reportUnusedVariable = "warning"
reportUnusedExpression = "warning"
